添加发布产品产生保全链功能

+++ src/main/java/com/baoquan/shuxin/dao/user/UserDao.java
@@ -21,7 +21,7 @@ public interface UserDao {
 
     void deleteUser(Long id);
 
-    User findByIdUserInfo(Long id);
+    User findByIdUserInfo(Integer id);
 
     User queryUserBalance(Long id);
